refactor: Car 객체 분리 및 메소드 추가
