+++ src/index.js
@@ -13,17 +13,6 @@ import {
 } from './utils.js';
 import { ERROR_MESSAGE } from './constants.js';
 
-export default class Car {
-  constructor(name) {
-    this.name = name;
-    this.forwardCount = 0;
-  }
-
-  generateRandomNumber = () => {
-    return MissionUtils.Random.pickNumberInRange(0, 9);
-  };
-}
-
 class CarRacingGame {
   constructor() {
     this.initElements();
